requirements.txt
streamlit
pandas
sqlalchemy
google.generativeai
python-dotenv